restructure into two apps for readability
